LayoutTests/imported/w3c/web-platform-tests/trusted-types/trusted-types-reporting-for-Document-write-expected.txt

CONSOLE MESSAGE: Refused to connect to http://localhost:8800/common/blank.html because it does not appear in the connect-src directive of the Content Security Policy.
CONSOLE MESSAGE: This requires a TrustedHTML value else it violates the following Content Security Policy directive: "require-trusted-types-for 'script'"
CONSOLE MESSAGE: Refused to connect to http://localhost:8800/common/blank.html because it does not appear in the connect-src directive of the Content Security Policy.
CONSOLE MESSAGE: Refused to connect to http://localhost:8800/common/blank.html because it does not appear in the connect-src directive of the Content Security Policy.
CONSOLE MESSAGE: This requires a TrustedHTML value else it violates the following Content Security Policy directive: "require-trusted-types-for 'script'"
CONSOLE MESSAGE: Refused to connect to http://localhost:8800/common/blank.html because it does not appear in the connect-src directive of the Content Security Policy.

PASS No violation reported for write() with TrustedHTML arguments.
PASS Violation report for plain string for write() with at least one plain string argument.
PASS No violation reported for writeln() with TrustedHTML arguments.
PASS Violation report for plain string for writeln() with at least one plain string argument.
